requirements.txt
# FastAPI及相关
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Redis和Celery
redis==5.0.1
celery==5.3.4
kombu==5.3.4

# 认证和安全
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# HTTP客户端
httpx==0.25.2
aiohttp==3.9.1

# 系统监控
psutil==5.9.6
monitorcontrol==3.0

# 工具库
python-dotenv==1.0.0
PyYAML==6.0.1
pandas==2.1.4
numpy==1.25.2

# 文件处理
python-magic==0.4.27
Pillow==10.1.0

# MinIO对象存储
minio==7.2.0

# 日志
structlog==23.2.0
colorlog==6.8.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# 其他
schedule==1.2.0
jinja2==3.1.2